+++ Model/Adminhtml/Source/Country.php
@@ -2,6 +2,12 @@
 
 namespace PlacetoPay\Payments\Model\Adminhtml\Source;
 
+use PlacetoPay\Payments\Countries\BelizeCountryConfig;
+use PlacetoPay\Payments\Countries\ChileCountryConfig;
+use PlacetoPay\Payments\Countries\CountryConfig;
+use PlacetoPay\Payments\Countries\EcuadorCountryConfig;
+use PlacetoPay\Payments\Countries\HondurasCountryConfig;
+
 class Country
 {
     const COLOMBIA = 'CO';
@@ -9,6 +15,16 @@ class Country
     const ECUADOR = 'EC';
     const CHILE = 'CL';
     const PUERTO_RICO = 'PR';
+    const HONDURAS = 'HN';
+    const BELIZE = 'BZ';
+
+    public const COUNTRIES_CONFIG = [
+        EcuadorCountryConfig::class,
+        ChileCountryConfig::class,
+        HondurasCountryConfig::class,
+        BelizeCountryConfig::class,
+        CountryConfig::class
+    ];
 
     public function toOptionArray(): array
     {
@@ -33,6 +49,14 @@ class Country
                 'value' => self::PUERTO_RICO,
                 'label' => __('Puerto Rico'),
             ],
+            [
+                'value' => self::HONDURAS,
+                'label' => __('Honduras'),
+            ],
+            [
+                'value' => self::BELIZE,
+                'label' => __('Belize')
+            ]
         ];
     }
 }