+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Honduras and Belize endpoints
+
 ## [1.8.11] - 2022-05-09
 
 ### Added